move suite generator to own module, break suite_xml package
