+++ corehq/apps/app_manager/management/commands/find_broken_builds.py
@@ -5,6 +5,7 @@ from django.core.management.base import CommandError
 from corehq.apps.app_manager import suite_xml
 from corehq.apps.app_manager.exceptions import SuiteValidationError
 from corehq.apps.app_manager.models import Application
+from corehq.apps.app_manager.suite_xml.utils import validate_suite
 from corehq.apps.app_manager.xform import XForm
 from dimagi.utils.couch.database import iter_docs
 
@@ -49,7 +50,7 @@ def broken_suite_files(build):
         error = 'build has no attachment files/suite.xml'
     else:
         try:
-            suite_xml.validate_suite(suite)
+            validate_suite(suite)
         except SuiteValidationError as error:
             pass
     if error:
