+++ corehq/apps/app_manager/suite_xml/post_process/instances.py
@@ -1,7 +1,7 @@
 from collections import defaultdict
 import re
 from corehq.apps.app_manager.exceptions import UnknownInstanceError
-from corehq.apps.app_manager.suite_xml.generator import SuiteContributor
+from corehq.apps.app_manager.suite_xml.contributors import SuiteContributor
 from corehq.apps.app_manager.suite_xml.xml_models import Instance
 from dimagi.utils.decorators.memoized import memoized
