+++ corehq/apps/app_manager/suite_xml/utils.py
@@ -1,3 +1,8 @@
+from lxml import etree
+from corehq.apps.app_manager.exceptions import SuiteValidationError
+from corehq.apps.app_manager.suite_xml.xml_models import Suite
+
+
 def get_select_chain(app, module, include_self=True):
         select_chain = [module] if include_self else []
         current_module = module
@@ -30,3 +35,24 @@ def get_select_chain_meta(app, module):
         }
         for i, mod in reversed(list(enumerate(select_chain)))
     ]
+
+
+def validate_suite(suite):
+    if isinstance(suite, unicode):
+        suite = suite.encode('utf8')
+    if isinstance(suite, str):
+        suite = etree.fromstring(suite)
+    if isinstance(suite, etree._Element):
+        suite = Suite(suite)
+    assert isinstance(suite, Suite),\
+        'Could not convert suite to a Suite XmlObject: %r' % suite
+
+    def is_unique_list(things):
+        return len(set(things)) == len(things)
+
+    for detail in suite.details:
+        orders = [field.sort_node.order for field in detail.fields
+                  if field and field.sort_node]
+        if not is_unique_list(orders):
+            raise SuiteValidationError('field/sort/@order must be unique per detail')
+
